move torsion test helpers into utils.geometry
